+++ src/main/java/com/ktds/dsquare/board/qna/dto/BriefQuestionResponse.java
@@ -30,11 +30,11 @@ public class BriefQuestionResponse {
     private Boolean managerAnswerYn;
     private Boolean atcYn;
     private List<String> tags;
-    private Integer likeCnt;
+    private Long likeCnt;
     private Boolean likeYn;
 
     public static BriefQuestionResponse toDto(Question question, MemberInfo writerInfo, CategoryResponse category, Long answerCnt,
-                                              Boolean managerAnswerYn, Integer likeCnt, Boolean likeYn, Long commentCnt){
+                                              Boolean managerAnswerYn, Long likeCnt, Boolean likeYn, Long commentCnt){
         List<QuestionTag> questionTags = question.getQuestionTags();
         List<String> tags = new ArrayList<>();
         for(QuestionTag questionTag : questionTags)
